feat(web): complete issue 172 route boundaries and loading shells
